Added Website pages
Login is not secure, the web server handles this i think, also I could not get permission from the web browser to look at result.txt on my local machine however a web server likely will not have an issue. Functionality of upload of problems and pdf needs to be added and thats about it!